pyproject.toml
[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[project]
name = "mempp"
version = "0.1.0"
description = "AI agent memory system for learning reusable procedures from task execution traces"
readme = "README.md"
authors = [
    { name = "Koichi Shiraishi", email = "zchee.io@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "cachetools>=5.5.2",
    "numpy>=2.3.2",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "structlog>=25.4.0",
    "tenacity>=9.1.2",
    # type stubs helpful for dev
    "types-cachetools>=6.2.0.20250827",
    "types-pyyaml>=6.0.12.20250822",
]

[project.scripts]
mempp = "mempp:main"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pyright>=1.1.405",
    "ruff>=0.12.12",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
]

[project.optional-dependencies]
# Install feature sets as needed, e.g. `uv add .[api,heavy,pinecone]`
api = ["fastapi>=0.116.1", "uvicorn>=0.35.0", "prometheus-client>=0.22.1", "aiokafka>=0.12.0", "psutil>=7.0.0"]
openai = ["openai>=1.106.1"]
anthropic = ["anthropic>=0.66.0"]
gemini = ["google-genai>=1.33.0"]
ray = ["ray>=2.49.1"]
pinecone = ["pinecone>=7.3.0", "pinecone-text>=0.11.0"]
heavy = [
  "torch>=2.8.0",
  "transformers>=4.56.1",
  "sentence-transformers>=5.1.0",
  "spacy>=3.8.7",
  "scikit-learn>=1.7.1",
  "yake>=0.6.0",
  "keybert>=0.9.0",
  "pandas>=2.3.2",
  "scipy>=1.16.1",
  "wandb>=0.21.3",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true

# See https://docs.astral.sh/ruff/rules for more information.
select = [
  "ARG",    # flake8-unused-arguments
  "B",      # flake8-bugbear
  "C4",     # flake8-comprehensions
  "D101",   # pydocstyle/undocumented-public-class
  "D102",   # pydocstyle/undocumented-public-method
  "D103",   # pydocstyle/undocumented-public-function
  "E",      # pycodestyle (Error)
  "F",      # Pyflakes
  "I",      # isort
  "INP001", # flake8-no-pep420/implicit-namespace-package
  "N",      # pep8-naming
  "SIM",    # flake8-simplify
  "TD002",  # flake8-todos/missing-todo-author
  "TD006",  # flake8-todos/invalid-todo-capitalization
  "TID",    # flake8-tidy-imports
  "UP",     # pyupgrade
  "W",      # pycodestyle (Warning)
]

ignore = [
  "N802", # Function name {name} should be lowercase
  "N805", # First argument of a method should be named self

  # ruff formatter recommends to disable those. see https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
  "COM812", # Trailing comma missing
  "COM819", # Trailing comma prohibited
  "D206",   # Docstring should be indented with spaces, not tabs
  "D300",   # Use triple double quotes """
  "E111",   # Indentation is not a multiple of
  "E114",   # Indentation is not a multiple of {indent_size} (comment)
  "E117",   # Over-indented (comment)
  "E203",   # Whitespace before '{symbol}'
  "E501",   # Line too long ({width} > {limit})
  "ISC001", # Implicitly concatenated string literals on one line
  "ISC002", # Implicitly concatenated string literals over multiple lines
  "Q000",   # Single quotes found but double quotes preferred
  "Q001",   # Single quote multiline found but double quotes preferred
  "Q002",   # Single quote docstring found but double quotes preferred
  "Q003",   # Change outer quotes to avoid escaping inner quotes
  "W191",   # Indentation contains tabs
  # Relax docstring/import/style until repo fully documented
  "D101", "D102", "D103",
  "I001",
  "B903",
  "SIM102", "SIM105",
  "W293",
  "N814", "N817",
  "C419",
  "ARG002",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["D101", "D102", "D103"]

[tool.ruff.format]
preview = true

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
check_untyped_defs = true
strict = true
disable_error_code = [
  "no-untyped-call", # Call to untyped function "xxx" in typed context
  "no-any-return",   # Returning Any from function declared to return "xxx"
  "misc",            # Class cannot subclass "xxx" (has type "Any")
  "unused-ignore",   # Unused "type: ignore" comment
]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"
